pyproject.toml
[project]
name = "vneurotk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "cebra==0.6.0",
    "h5py>=3.16.0",
    "httpx[socks]>=0.28.1",
    "loguru>=0.7.3",
    "mne>=1.11.0",
    "mne-bids>=0.17.0",
    "numpy",
    "pandas>=2.3.3",
    "pillow",
    "sentencepiece>=0.2.1",
    "thingsvision>=1.4.4",
    "timm",
    "torch>=2.4.1",
    "transformers",
    "trialcebra>=0.0.1a2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/vneurotk"]

[dependency-groups]
dev = [
    "hatchling>=1.29.0",
    "ipykernel>=7.2.0",
    "pytest>=9.0.2",
]
